Pruned the Authorization features and updated the user, role and permission table
