--- src/modules/auth/decorators/roles.decorator.ts
@@ -1,6 +1,0 @@
-import { SetMetadata } from '@nestjs/common';
-import { Role } from '../enum';
-
-export const ROLES_KEY = 'roles';
-export const Roles = (...roles: [Role, ...Role[]]) =>
-  SetMetadata('roles', roles);
