--- src/modules/auth/enum/index.ts
@@ -1,1 +1,0 @@
-export * from './roles.enum';
