--- src/modules/auth/enum/roles.enum.ts
@@ -1,4 +1,0 @@
-export enum Role {
-  ADMIN = 'ADMIN',
-  USER = 'USER',
-}
