--- src/modules/auth/guards/index.ts
@@ -1,2 +1,0 @@
-export * from './jwt.guard';
-export * from './refresh-token.guard';
