--- src/modules/auth/guards/roles/index.ts
@@ -1,1 +1,0 @@
-export * from './roles.guard';
