--- src/modules/auth/guards/roles/roles.guard.ts
@@ -1,22 +1,0 @@
-import { CanActivate, ExecutionContext, Injectable } from '@nestjs/common';
-import { Reflector } from '@nestjs/core';
-import { Observable } from 'rxjs';
-import { Role } from '../../enum';
-import { ROLES_KEY } from '../../decorators/roles.decorator';
-
-@Injectable()
-export class RolesGuard implements CanActivate {
-  constructor(private reflector: Reflector) {}
-
-  canActivate(context: ExecutionContext): boolean {
-    const requiredRoles = this.reflector.getAllAndOverride<Role[]>(ROLES_KEY, [
-      context.getHandler(),
-      context.getClass(),
-    ]);
-
-    const user = context.switchToHttp().getRequest().user;
-    console.log(user);
-    const checkRequiredRoles = requiredRoles.some((role) => user.role === role);
-    return checkRequiredRoles;
-  }
-}
