+++ src/modules/auth/index.ts
@@ -1,4 +1,3 @@
 export * from './auth.module';
 export * from './controller';
 export * from './service';
-export * from './enum';
