--- src/modules/auth/types/current-user.d.ts
@@ -1,6 +1,0 @@
-import { Role } from '../enum';
-
-export type CurrentUser = {
-  id: number;
-  role: Role;
-};
