--- src/modules/db-module/entities/authorization.ts
@@ -1,33 +1,0 @@
-import { Action, Resource } from 'src/shared/utils/enums';
-import {
-  Column,
-  Entity,
-  JoinTable,
-  ManyToMany,
-  PrimaryGeneratedColumn,
-} from 'typeorm';
-
-@Entity({ name: 'permission' })
-export class Permission {
-  @PrimaryGeneratedColumn()
-  id: number;
-
-  @Column({ type: 'enum', enum: Resource })
-  resource: Resource;
-
-  @Column({ type: 'enum', enum: Action, array: true })
-  actions: Action;
-}
-
-@Entity({ name: 'role' })
-export class Role {
-  @PrimaryGeneratedColumn()
-  id: number;
-
-  @Column({ unique: true })
-  name: string;
-
-  @ManyToMany(() => Permission, { cascade: true })
-  @JoinTable()
-  permissions: Permission[];
-}
