+++ src/modules/db-module/entities/index.ts
@@ -1,2 +1,4 @@
 export * from './movies';
 export * from './users';
+export * from './role.entity';
+export * from './permission.entity'
