--- src/modules/roles/index.ts
@@ -1,3 +1,0 @@
-export * from './controller';
-export * from './service';
-export * from './roles.module';
