--- src/shared/decorators/index.ts
@@ -1,1 +1,0 @@
-export * from './permissions.decorator';
