--- src/shared/decorators/permissions.decorator.ts
@@ -1,5 +1,0 @@
-import { SetMetadata } from '@nestjs/common';
-import { Permission } from 'src/modules/roles/dto/create-role.dto';
-
-export const Permissions = (permissions: Permission[]) =>
-  SetMetadata('permissions', permissions);
