--- src/shared/utils/enums/action.enum.ts
@@ -1,8 +1,0 @@
-export enum Action {
-  READ = 'read',
-  CREATE = 'create',
-  UPDATE = 'update',
-  DELETE = 'delete',
-  ADD = 'add',
-  REMOVE = 'remove',
-}
