--- src/shared/utils/enums/index.ts
@@ -1,2 +1,0 @@
-export * from './action.enum';
-export * from './resource.enum';
