--- src/shared/utils/enums/resource.enum.ts
@@ -1,4 +1,0 @@
-export enum Resource {
-  MOVIES = 'movies',
-  USERS = 'users',
-}
